% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitARCO.R
\name{fitArCo}
\alias{fitArCo}
\title{Estimates the ArCo using the model selected by the user}
\usage{
fitARCO(data, fn, p.fn, treated.unity, t0, lag = 0, Xreg = NULL,
  display = TRUE, HACweights = 1, alpha = 0.05, ...)
}
\arguments{
\item{data}{A list of matrixes or dataframes of length q. Each matrix is T X n and it contains observations of a single variable for all units and all periods of time. Even in the case of a single variable (q=1), the matrix must be inside a list.}

\item{fn}{The function used to estimate the first stage model. This function must receive only two arguments in the following order: X (independent variables), y (dependent variable). If the model requires additional arguments they must be supplied inside the function fn.}

\item{p.fn}{The function used to estimate the predict using the first stage model. This function also must receive only two arguments in the following order: model (model estimated in the first stage), newdata (out of sample data to estimate the second stage). If the prediction requires additional arguments they must be supplied inside the function p.fn.}

\item{treated.unity}{Single number indicating the unity where the intervention took place.}

\item{t0}{Single number indicating the intervention period.}

\item{lag}{Number of lags in the first stage model. Default is 0, i.e. only contemporaneous variables are used.}

\item{Xreg}{Exogenous controls.}

\item{display}{Default is TRUE. Shows the counterfactual plot automaticaly.}

\item{HACweights}{Vector of weights for the robust covariance matrix of the delta statistics. Default is 1 for the lag 0 and 0 for all other lags.}

\item{alpha}{Significance level for the delta.}

\item{...}{Aditional parameters used only if display = TRUE.}
}
\description{
This description may be useful to clarify the notation and understand how the arguments must be supplied to the functions. 
* units: Each unity is indexed by a number between 1,...,n. They are for exemple: countries, states, municipalities, firms, etc. 
* Variables:  For each unity and for every time period t=1,...,T we observe q_i >= 1 variables. They are for example: GDP, inflation, sales, etc.
* Intervention:  The intervention took place only in the treated unity at time t0=L0*T, where L0 is in (0,1).
}
\examples{
#############################
data(data.q1) # = First unity was treated on t=51 by adding a constant equal 3
data=list(data.q1) # = Even if q=1 the data must be in a list
## == Fitting the ArCo using linear regression == ##
# = creating fn and p.fn function = #
fn=function(X,y){
return(lm(y~X))
}
p.fn=function(model,newdata){
b=coef(model)
return(cbind(1,newdata) \%*\% b)}
ArCo=fitArCo(data = data,fn = fn, p.fn = p.fn, treated.unity = 1 , t0 = 51)

#############################

# = First unity was treated on t=51 by adding constants 3 and -3 for the first and second variables
data(data.q2) # data is already a list
## == Fitting the ArCo using the package randomForest == ##
require(randomForest)
## == Bartlett kernel weights for two lags == ##
l=2
w <- seq(1, 0, by = -(1/(l + 1)))[1:(l+1)]
ArCo2=fitArCo(data = data.q2,fn = randomForest, p.fn = predict,
treated.unity = 1 , t0 = 51, HACweights = w)
## == Fitting the ARCO using the package glmnet via LASSO and crossvalidation == ##
require(glmnet)
## == Bartlett kernel weights for two lags == ##
ARCO3=fitArCo(data = data.q2,fn = cv.glmnet, p.fn = predict, 
treated.unity = 1 , t0 = 51, HACweights = w)

}
\references{
Carvalho, C., Masini, R., Medeiros, M. (2016) "ArCo: An Artificial Counterfactual Approach For High-Dimensional Panel Time-Series Data.".
}
\keyword{ArCo}
