% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{inflationNFP}
\alias{inflationNFP}
\title{Dataset used on the empirical example by Carvalho, Masini and Medeiros (2016).}
\format{A list with two matrixes of 56 rows and 9 variables.}
\usage{
data(inflationNFP)
}
\description{
This is the data from the \emph{nota fiscal paulista} (NFP) example from Carvalho, Masini and Medeiros (2016). The variables are the food away from home component of the inflation and the GDP for 9 metropolitan areas in Brazil. Each variable is represented by a matrix inside the list. The treated unit is the Sao Paulo metropolitan area, which is the first column in each matrix. The treatment took place at \eqn{t_0=34}.
}
\references{
Carvalho, C., Masini, R., Medeiros, M. (2016) "ArCo: An Artificial Counterfactual Approach For High-Dimensional Panel Time-Series Data.".
}
\keyword{datasets}
