% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fitArCo.R
\name{plot.fitArCo}
\alias{plot.fitArCo}
\title{Plots realized values and the counter factual estimated by the fitArCo function}
\usage{
\method{plot}{fitArCo}(x, ylab = NULL, main = NULL, plot = NULL,
  ncol = 1, display.fitted = FALSE, y.min = NULL, y.max = NULL, ...)
}
\arguments{
\item{x}{An ArCo object estimated using the fitArCo function.}

\item{ylab}{n dimensional character vector, where n is the length of the plot argument.}

\item{main}{n dimensional character vector, where n is the length of the plot argument.}

\item{plot}{n dimensional numeric vector where each element represents an ArCo unity. If NULL, all unities will be plotted and ylab. If, for example, plot=c(1,2,5) only unities 1 2 and 5 will be plotted according to the order specified by the user on the fitArCo.}

\item{ncol}{Number of columns when multiple plots are displayed.}

\item{display.fitted}{If TRUE the fitted values of the estimation window are also plotted (default=FALSE).}

\item{y.min}{n dimensional numeric vector defining the lower bound for the y axis.}

\item{y.max}{n dimensional numeric vector defining the upper bound for the y axis.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots realized values and the counter factual estimated by the fitArCo function
}
