% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.q1}
\alias{data.q1}
\title{A dataset used in the examples}
\format{A matrix with 100 rows and 20 variables}
\usage{
data(data.q1)
}
\description{
A dataset used in the examples
}
\keyword{datasets}
