% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data.q2}
\alias{data.q2}
\title{A dataset used in the examples}
\format{A list with 2 matrixes of 100 rows and 20 variables}
\usage{
data(data.q2)
}
\description{
A dataset used in the examples
}
\keyword{datasets}
